Fixed auto-updater, added 32-bit support
Whoops. Turns out that grep really, really cares about case sensitivity; "steamapps" was spelled as "SteamApps" before, which broke it entirely and I didn't even know.

On another note, I decided to use this opportunity to make the script automatically detect whether you're running a 64-bit operating system or a 32-bit operating system. It does this by checking if "Program Files (x86)" exists. If you, for some stupid reason, are running a 32-bit OS and have a folder called that in the root of your C: drive, then it'll mess up and install there instead. So good job. To fix that, just get rid of that folder.

If anyone experiences issues with this, then tell me. For reference, though, this iteration of the script doesn't quite support custom installation paths. So to make it easier, I've added an updater-legacy.bat. If you have Team Fortress 2 installed somewhere other than the default location, use that script and replace the path in it with your installation path, in the same format as the path already in that file.